splitting and updating tests for benchmarks and tdt
